Combined imd and tmd into md as single module and added tsconfig
