+++ tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "compilerOptions": {
+        "lib": [
+            "DOM"
+        ],
+        "module": "commonjs",
+        "esModuleInterop": true,
+        "forceConsistentCasingInFileNames": true,
+        "strict": true,
+        "skipLibCheck": true
+    }
+}
